# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AI Alpha API",
    "description": "API documentation for AI Alpha",
    "version": "2.1"
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/chart": {
      "post": {
        "tags": [
          "Chart"
        ],
        "summary": "Adds a new support and resistance lines record for a coin.",
        "description": "Adds a new support and resistance lines record for a coin.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "coin_id": {
                  "type": "integer"
                },
                "pair": {
                  "type": "string"
                },
                "temporality": {
                  "type": "string"
                },
                "coin_name": {
                  "type": "string"
                },
                "support_1": {
                  "type": "number"
                },
                "support_2": {
                  "type": "number"
                },
                "support_3": {
                  "type": "number"
                },
                "support_4": {
                  "type": "number"
                },
                "resistance_1": {
                  "type": "number"
                },
                "resistance_2": {
                  "type": "number"
                },
                "resistance_3": {
                  "type": "number"
                },
                "resistance_4": {
                  "type": "number"
                }
              },
              "required": [
                "coin_id",
                "pair",
                "temporality",
                "coin_name",
                "support_1",
                "support_2",
                "support_3",
                "support_4",
                "resistance_1",
                "resistance_2",
                "resistance_3",
                "resistance_4"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "New chart record created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Chart"
        ],
        "summary": "Get the most recent support and resistance lines of a requested coin.",
        "description": "Get the most recent support and resistance lines of a requested coin.",
        "parameters": [
          {
            "name": "coin_name",
            "in": "query",
            "description": "",
            "required": false,
            "type": "string",
            "schema": {}
          },
          {
            "name": "coin_id",
            "in": "query",
            "description": "",
            "required": false,
            "type": "integer",
            "schema": {}
          },
          {
            "name": "temporality",
            "in": "query",
            "description": "",
            "required": true,
            "type": "string",
            "schema": {}
          },
          {
            "name": "pair",
            "in": "query",
            "description": "",
            "required": true,
            "type": "string",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved chart values",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "object"
                },
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/chart/total3": {
      "get": {
        "tags": [
          "Chart"
        ],
        "summary": "Retrieve and calculate total market cap data for the top 3 cryptocurrencies.",
        "description": "Retrieve and calculate total market cap data for the top 3 cryptocurrencies.",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "",
            "required": false,
            "type": "integer",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved data",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "status": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}